requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.0
pillow>=8.0.0